feat: check if file exists before calling the API

+++ extract_parameters.py
@@ -1,5 +1,6 @@
 from openai import OpenAI
 import concurrent.futures
+import os  # Import the os module
 
 client = OpenAI()
 
@@ -10,20 +11,27 @@ with open("signs_splited.out", "r") as file:
 
 
 def make_api_call(text, index):
-    response = client.chat.completions.create(
-        model="ft:gpt-3.5-turbo-1106:personal::8pdI51i3",
-        messages=[
-            {"role": "system", "content": system},
-            {"role": "user", "content": text},
-        ],
-        temperature=0.0,
-        top_p=1,
-    )
     filename = f"tmp/response_{index}.txt"
-    with open(filename, "w") as file:
-        file.write(
-            text + "\n\n" + "```json\n" + response.choices[0].message.content + "\n```"
+    if not os.path.exists(filename):
+        response = client.chat.completions.create(
+            model="ft:gpt-3.5-turbo-1106:personal::8pdI51i3",
+            messages=[
+                {"role": "system", "content": system},
+                {"role": "user", "content": text},
+            ],
+            temperature=0.0,
+            top_p=1,
         )
+        with open(filename, "w") as file:
+            file.write(
+                text
+                + "\n\n"
+                + "```json\n"
+                + response.choices[0].message.content
+                + "\n```"
+            )
+    else:
+        print(f"{filename} already exists. Skipping API call.")
     return filename
 
 
@@ -34,4 +42,5 @@ with concurrent.futures.ThreadPoolExecutor(max_workers=10) as executor:
     }
     for future in concurrent.futures.as_completed(future_to_text):
         filename = future.result()
-        print(f"Output written to {filename}")
+        if os.path.exists(filename):
+            print(f"Output written to {filename}")
